requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
pandas==2.2.3
plotly==5.17.0
gunicorn==20.1.0
dash==2.14.1
dash-auth==2.3.0